requirements.txt
Django~=4.1.3
reportlab~=3.6.12
django-easy-pdf>=0.2.0.dev0
requests~=2.28.1
xhtml2pdf~=0.2.8
djangorestframework~=3.14.0
WeasyPrint>=0.34